+++ src_files/src/src_cliente/client_comandos/cmd_enemigo_despliega_unidad.h
@@ -9,7 +9,7 @@ private:
     uint8_t id_jugador;
     uint8_t id_unidad;
     uint8_t tipo_unidad;
-    long tiempo_entrenamiento;
+    uint16_t tiempo_entrenamiento;
     Coordenadas& coords_spawn;
 
 public:
