+++ src_files/src/src_server/server_comandos/cmd_mover_unidad.cpp
@@ -1,6 +1,6 @@
 #include "cmd_mover_unidad.h"
 
-CmdMoverUnidadServer::CmdMoverUnidadServer(uint8_t id_unidad, char direccion, long tiempo) :
+CmdMoverUnidadServer::CmdMoverUnidadServer(uint8_t id_unidad, char direccion, uint16_t tiempo) :
                                                   id_unidad(id_unidad),
                                                   direccion(direccion),
                                                   tiempo(tiempo){}
