+++ src_files/src/src_server/server_game/server_unidades/sardaukar.cpp
@@ -7,7 +7,7 @@ Sardaukar::Sardaukar(uint8_t id, Jugador& duenio, Mapa& mapa, YAML::Node& atribu
     this->armas = atributos_unidad["Infanteria"]["Sardaukar"]["Arma"].as<std::vector<std::string>>();
     this->rango = atributos_unidad["Infanteria"]["Sardaukar"]["Rango"].as<uint8_t>();
     this->velocidad = atributos_unidad["Infanteria"]["Sardaukar"]["Velocidad"].as<int16_t>();
-    this->tiempo_entrenamiento = atributos_unidad["Infanteria"]["Sardaukar"]["TiempoEntrenamiento"].as<float>();
+    this->tiempo_entrenamiento = atributos_unidad["Infanteria"]["Sardaukar"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Infanteria"]["Sardaukar"]["Vida"].as<int16_t>();
     this->costo = atributos_unidad["Infanteria"]["Sardaukar"]["Costo"].as<uint8_t>();
