+++ src_files/src/src_server/server_game/server_unidades/tanque.cpp
@@ -7,7 +7,7 @@ Tanque::Tanque(uint8_t id, Jugador& duenio, Mapa& mapa, YAML::Node& atributos_un
     this->armas = atributos_unidad["Vehiculo"]["Tanque"]["Arma"].as<std::vector<std::string>>();
     this->rango = atributos_unidad["Vehiculo"]["Tanque"]["Rango"].as<uint8_t>();
     this->velocidad = atributos_unidad["Vehiculo"]["Tanque"]["Velocidad"].as<int16_t>();
-    this->tiempo_entrenamiento = atributos_unidad["Vehiculo"]["Tanque"]["TiempoEntrenamiento"].as<float>();
+    this->tiempo_entrenamiento = atributos_unidad["Vehiculo"]["Tanque"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Vehiculo"]["Tanque"]["Vida"].as<int16_t>();
     this->costo = atributos_unidad["Vehiculo"]["Tanque"]["Costo"].as<uint8_t>();
